requirements.txt
pdfplumber==0.11.5
PyMuPDF
pdf2image
Pillow
opencv-python
pytesseract
numpy